Bind httpServer to specific ip and udp socket to ssdp address
This change allows to run a http server for each ip address of the
device, making upnp to work even with multiple network interfaces.
Binding to 0.0.0.0 does not work because this address is reported back
in the ssdp packet as the address to use to retrieve the xml, witch is
incorrect.